% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catego_average.R
\name{catego_average}
\alias{catego_average}
\title{Mean values of categories.}
\usage{
catego_average(data_in, data_meta)
}
\arguments{
\item{data_in}{....}

\item{stat_meta}{....}
}
\value{
....
}
\description{
Function to get mean values of elevatio categories.
}
\examples{
mea_na(c(2 ,3, NA))
mea_na(c(NA, NA, NA))
}
