% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fac2num.R
\name{fac2num}
\alias{fac2num}
\title{Factor values to numeric.}
\usage{
fac2num(x)
}
\arguments{
\item{x}{factor value(s).}
}
\value{
The numeric value of the factor.
}
\description{
Function to transform a factor to approximately its original numeric value.
}
\examples{
f <- factor(sample(runif(5), 20, replace = TRUE))
fac2num(f)
}
