% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idaweb_data_prep.R
\name{idaweb_data_prep}
\alias{idaweb_data_prep}
\title{MeteoSwiss Idaweb climate data preparation.}
\usage{
idaweb_data_prep(param_code, order_dir, out_dir, start_day = "1981-01-01",
  end_day = "2017-12-31")
}
\arguments{
\item{param_code}{Parameter code from climatological variable as downloaded
from Idaweb (e.g. "tre200d0" or "gre000d0")}

\item{order_dir}{...}

\item{out_dir}{...}

\item{start_day}{...}

\item{end_day}{...}
}
\value{
The arithmetic mean of values in \code{x}.
}
\description{
Preparation of raw climatological data obtained from national weather and
climate service of Switzerland (MeteoSwiss) via Idaweb.
}
\examples{
mea_na(c(2 ,3, NA))
mea_na(c(NA, NA, NA))
}
