% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_cycl_elev.R
\name{image_cycle_elev}
\alias{image_cycle_elev}
\title{Image plots yearly cycles of trend magnitudes.}
\usage{
image_cycle_elev(data_in, stat_meta, colors, breaks, main = "", unit = "",
  margins_1 = c(1.6, 2.5, 1.6, 0), margins_2 = c(1.6, 0.5, 1.6, 1.7),
  isotherm_data = NULL, add_isotherms = FALSE, pos_iso_text = NULL,
  pos_ylab_lal = c(0.27, 0.71, 0.985))
}
\arguments{
\item{data_in}{...}

\item{data_mk}{...}

\item{data_in_me}{...}

\item{data_meta}{...}

\item{window_width}{...}

\item{cover_thresh}{...}

\item{method_analys}{...}
}
\value{
...
}
\description{
...
}
\examples{

}
