% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_scale.R
\name{image_scale}
\alias{image_scale}
\title{Scale for  image plot.}
\usage{
image_scale(z, zlim, col = heat.colors(12), breaks, horiz = TRUE,
  ylim = NULL, xlim = NULL, ...)
}
\arguments{
\item{z}{...}

\item{zlim}{...}

\item{col}{...}

\item{breaks}{...}

\item{horiz}{...}

\item{ylim}{...}

\item{xlim}{...}
}
\value{
...
}
\description{
...
}
\examples{

}
