% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iso_days.R
\name{iso_days}
\alias{iso_days}
\title{Determination of temperature isotherms.}
\usage{
iso_days(data_in, isotherm = 0, meta_stat, smoo_val = 0.75)
}
\arguments{
\item{data_in}{Input temperature data after moving average statistics to get
mean yearly cycles of temperature of stations.}

\item{isotherm}{What isotherm to calculate (e.g. 0 for 0°C isotherm)}

\item{meta_stat}{Station meta information.}

\item{smoo_val}{Smooth values using LOESS (f = smoo_val)}
}
\value{
Days of the year which closest to selected isotherm.
}
\description{
Calculation of temperature isotherms for image plot (see image_cycle_elev())
based on temperature data (usually after applying moving average filter).
}
