% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loess_NA_restore.R
\name{loess_NA_restore}
\alias{loess_NA_restore}
\title{Local Polynomial Regression Fitting.}
\usage{
loess_NA_restore(data_in, smoo_val = 0.2, NA_restore = TRUE)
}
\arguments{
\item{data_in}{...}

\item{smoo_val}{...}

\item{NA_restore}{...}
}
\value{
...
}
\description{
...
}
\examples{
pres2ahum(vapor_pres = 5.5, temp = 20)
}
