% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/order_add_stat.R
\name{order_add_stat}
\alias{order_add_stat}
\title{Order data set by altitude. Add missing stations.}
\usage{
order_add_stat(data_in, meta_stat)
}
\arguments{
\item{data_in}{Data frame after moving average trends analysis.}

\item{meta_stat}{Stations meta information.}
}
\value{
Data frame with columns ordered according to station elevation.
}
\description{
Orders columns according to elevatoin of stations and adds missing stations for
which no data available.
}
