% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cycl_elev.R
\name{plot_cycl_elev}
\alias{plot_cycl_elev}
\title{Plot yearly cycle and elevation dependency.}
\usage{
plot_cycl_elev(data_in, data_mk, data_in_me, data_meta, main_text = "",
  margins_1 = c(1.4, 1.8, 1.8, 0.2), margins_2 = c(1.4, 0.2, 1.8, 3.5),
  no_col = F, show_mk = F, aggr_cat_mean = F, with_hom_dat = F,
  smooth_val = 0.2, mk_sig_level = 0.05, add_st_num = T)
}
\arguments{
\item{data_in}{...}

\item{data_mk}{...}

\item{data_in_me}{...}

\item{data_meta}{...}

\item{window_width}{...}

\item{cover_thresh}{...}

\item{method_analys}{...}
}
\value{
...
}
\description{
...
}
\examples{
pres2ahum(vapor_pres = 5.5, temp = 20)
}
