% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pres2ahum.R
\name{pres2ahum}
\alias{pres2ahum}
\title{Water vapor pressure to absolute humidity.}
\usage{
pres2ahum(vapor_pres, temp, Rw = 461.5)
}
\arguments{
\item{vapor_pres}{Vapor pressure [hPa]}

\item{temp}{Ambient temperature [°C]}

\item{Rw}{Specific gas constant [J/(kg*K)]}
}
\value{
Absolute humidity of air [g/m³].
}
\description{
Calcuation of absolute air humidty based on measurements of water vapor pressure
and ambient temperature using formula derived from Ideal gas law.
Absolute humidity ahum = vapor_pres / (Rw * tmean) where vapor_pres
is the vapor pressure, tmean the ambient temperature and Rw is the specific gas
constant and for water equals 461.5 J/(kg*K).
}
\examples{

pres2ahum(vapor_pres = 5.5, temp = 20)
}
