% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_coverage.R
\name{stat_coverage}
\alias{stat_coverage}
\title{Remove columns with only NA.}
\usage{
stat_coverage(data_in)
}
\arguments{
\item{data_in}{data frame with numeric columns...}
}
\value{
...
}
\description{
Function to determine sum automatically handling NAs. Calculates
sum disregarding NAs, except when all values are NA.
}
\examples{
sum_na(c(2 ,3, NA))
sum_na(c(NA, NA, NA))
}
