% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_data.R
\docType{data}
\name{station_meta}
\alias{station_meta}
\title{Test data station meta information.}
\format{A data frame with 7 columns:
\describe{
\item{stn}{Station ID}
\item{name}{Station name}
\item{alt}{Altitude}
\item{lat}{Latitude}
\item{lon}{Longitude}
\item{category}{Elevation category (low, middle, or high)}
\item{data_qual}{quality level of data (homogenized or quality-checked)}
}}
\usage{
station_meta
}
\description{
A dataset of meta information of climatological stations
}
\keyword{datasets}
