% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_data.R
\docType{data}
\name{temperature}
\alias{temperature}
\title{Test data temperature time series.}
\format{A data frame with 7 columns (date column + six columns with stations data)
and 10958 rows. Temperature values were randomly generated.}
\usage{
temperature
}
\description{
A test dataset of temperature time series (radomly generated) of six station
}
\keyword{datasets}
