% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catego_average.R
\name{catego_average}
\alias{catego_average}
\title{Mean values of elevation categories.}
\usage{
catego_average(data_in, data_meta)
}
\arguments{
\item{data_in}{Input data frame; Station data sorted by column}

\item{stat_meta}{Table containing meta information on investigated stations}
}
\value{
Mean value of elevation categores (low, middle, high)
}
\description{
Function to get mean values of elevatio categories.
}
\examples{
catego_average(temperature[, -1], station_meta)
}
