% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_cycl_elev.R
\name{image_cycle_elev}
\alias{image_cycle_elev}
\title{Image plots yearly cycles of trend magnitudes.}
\usage{
image_cycle_elev(data_in, stat_meta, colors, breaks, main = "", unit = "",
  margins_1 = c(1.6, 2.5, 1.6, 0), margins_2 = c(1.6, 0.5, 1.6, 1.7),
  isotherm_data = NULL, add_isotherms = FALSE, pos_iso_text = NULL,
  pos_ylab_lal = c(0.27, 0.71, 0.985))
}
\arguments{
\item{data_in}{Data frame with results of moving average analysis.}

\item{colors}{Vector of colors for z-value.}

\item{breaks}{Color breaks in z-range.}

\item{main}{Plot title.}

\item{unit}{Unit of input data.}

\item{margins_1}{Margins of left plot panel.}

\item{margins_2}{Margins of right plot panel.}

\item{isotherm_data}{Data frame with isotherm data.}

\item{add_isotherms}{Plot isotherms yes or no (T or F)}

\item{pos_iso_text}{Position is text of isotherms in plot.}

\item{pos_ylab_lal}{Posistion  y-labels (HS, MS,LS)}

\item{data_meta}{Meta data file of stations.}
}
\description{
Image plot of results of moving average analysis. Stations ordered accoring to
elevation along y-axis.
}
\examples{

}
