% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iso_days.R
\name{iso_days}
\alias{iso_days}
\title{Determination of temperature isotherms.}
\usage{
iso_days(data_in, isotherm = 0, meta_stat, smoo_val = 0.75)
}
\arguments{
\item{data_in}{...}

\item{data_mk}{...}

\item{data_in_me}{...}

\item{data_meta}{...}

\item{window_width}{...}

\item{cover_thresh}{...}

\item{method_analys}{...}
}
\value{
...
}
\description{
...
}
\examples{
pres2ahum(vapor_pres = 5.5, temp = 20)
}
