% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method_analys.R
\name{mann_kendall}
\alias{mann_kendall}
\title{Calculate significance of trend using Mann Kendall trend test.}
\usage{
mann_kendall(data_in, cover_thresh = 0.9)
}
\arguments{
\item{data_in}{vector of input data}

\item{cover_thresh}{Trend only calculated when percentage of NAs in input vector
does not exceed cover_thresh (0-1)}
}
\value{
Nonlinear prewithened Sen´s slope trend.
}
\description{
Calculation of non-linear prewhitened ...
}
\examples{
data_mk <-  c(2, 3, 5, 6 ,8, 5, 8, 12, NA, NA)
mann_kendall(data_in, cover_thresh = 0.5)
mann_kendall(data_in, cover_thresh = 0.9)
}
