% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moving_analys.R
\name{moving_analys}
\alias{moving_analys}
\title{Moving average data analysis.}
\usage{
moving_analys(dates, values, start_year, end_year, window_width, cover_thresh,
  method_analys)
}
\arguments{
\item{dates}{...}

\item{values}{...}

\item{start_year}{...}

\item{end_year}{...}

\item{window_width}{...}

\item{cover_thresh}{...}

\item{method_analys}{...}
}
\value{
...
}
\description{
...
}
\examples{
pres2ahum(vapor_pres = 5.5, temp = 20)
}
