% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/order_add_stat.R
\name{order_add_stat}
\alias{order_add_stat}
\title{Order data set by altitude. Add missing stations.}
\usage{
order_add_stat(data_in, meta_stat)
}
\arguments{
\item{data_in}{...}

\item{meta_stat}{...}
}
\value{
...
}
\description{
...
}
\examples{

}
