% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cycl_elev.R
\name{plot_cycl_elev}
\alias{plot_cycl_elev}
\title{Plot yearly cycle and elevation dependency.}
\usage{
plot_cycl_elev(data_in, data_mk, data_in_me, data_meta, main_text = "",
  margins_1 = c(1.4, 1.8, 1.8, 0.2), margins_2 = c(1.4, 0.2, 1.8, 3.5),
  no_col = F, show_mk = F, aggr_cat_mean = F, with_hom_dat = F,
  smooth_val = 0.2, mk_sig_level = 0.05, add_st_num = T)
}
\arguments{
\item{data_in}{Input data frame holding results of moving average analysis.}

\item{data_mk}{Input analysis with results of Mann-Kendall trend test.}

\item{data_in_me}{Data frame with annual average values of stations.}

\item{data_meta}{Station meta data.}

\item{main_text}{Plot title}

\item{margins_1}{Figure margins of left plot panel}

\item{margins_2}{Figure margins of right plot panel.}

\item{no_col}{Display elevation categories with different colors (T or F)}

\item{aggr_cat_mean}{If TRUE, uasnge mean average instead of median for calculation af category averages}

\item{with_hom_dat}{Display stations with homogenized time series (T or F)}

\item{smooth_val}{Smoothing value for LOESS.}

\item{mk_sig_level}{Significance level alpha.}

\item{add_st_num}{Add number of stations of each category top right (T or F)}

\item{snow_mk}{Snow days with significant trends with yellow dots (T or F)}
}
\value{
Plot with yearly cycles (lines; left panel) of moving average analysis results and
annual averages (right panel)
}
\description{
Plot yearly cycles of moving average analysis and annual averages to visualize
elevation dependency for multiple stations of different elevation categories.
}
