% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_points.R
\name{plot_points}
\alias{plot_points}
\title{Plot points.}
\usage{
plot_points(data_in, data_meta, ID, no_col = FALSE, is_hom = FALSE)
}
\arguments{
\item{dates}{...}

\item{values}{...}

\item{start_year}{...}

\item{end_year}{...}

\item{window_width}{...}

\item{cover_thresh}{...}

\item{method_analys}{...}
}
\value{
...
}
\description{
...
}
\examples{
pres2ahum(vapor_pres = 5.5, temp = 20)
}
