% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method_analys.R
\name{sens_slope}
\alias{sens_slope}
\title{Calculate nonlinear prewhitened Sen´s slope trend.}
\usage{
sens_slope(data_in, cover_thresh = 0.9)
}
\arguments{
\item{data_in}{vector of input data}

\item{cover_thresh}{Trend only calculated when percentage of NAs in input vector
does not exceed cover_thresh (0-1)}
}
\value{
Nonlinear prewithened Sen´s slope trend.
}
\description{
Calculation of non-linear prewhitened (mehthod of Yue Pilon) Sen´s slope trend
using \link{zyp.trend.vector}.
}
\examples{
data_slope <- c(2, 3, 5, 6 ,8, 5, 8, 12, NA, NA)
sens_slope(data_in, cover_thresh = 0.5)
sens_slope(data_in, cover_thresh = 0.9)
}
