% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_coverage.R
\name{stat_coverage}
\alias{stat_coverage}
\title{Remove columns with only NA.}
\usage{
stat_coverage(data_in)
}
\arguments{
\item{data_in}{Input data frame with numerich columns.}
}
\value{
Data frame without only-NA-columns
}
\description{
Remove columns in data frame which onla have NA-values
}
